also move threshold code
